lunches missile (well more bomb at present with no AI)
Added in that it doesn't explode immediately via proximity. Added in extra byte for probability ship has ECM fitted and logic to drain power on ECM fire. Checks if there is enough energy to do it safe but factoring in damage etc it can still kill a ship if say player shoots it when very low on energy